requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-allauth-templates-bootstrap4==0.34.12
django-annoying==0.10.6
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-jsonstore==0.4.1
django-material==1.6.3
django-model-utils==4.0.0
django-viewflow==1.6.0
idna==2.9
invoke==1.4.1
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
